修复关注用户/粉丝页面不能正常打开的问题 (#1516)

+++ src/Models/Models.BiliBili/User/RelatedUserResponse.cs
@@ -34,7 +34,7 @@ namespace Bili.Models.BiliBili
         /// 用户ID.
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "mid", Required = Required.Default)]
-        public int Mid { get; set; }
+        public long Mid { get; set; }
 
         /// <summary>
         /// 关注方式，0-未关注，2-已关注，3-已互粉.
